refactor and improve documentation
